# openapi.yaml
openapi: 3.0.3
info:
  title: rustdesk-panel-api
  description: >-
    RustDesk 管理平台后端契约（单一契约源，oapi-codegen design-first）。
    M1 交付 healthz + 认证域 + 用户自身端点共 25 个 operation；
    M2 交付核心业务域 43 个 operation（设备端协议 heartbeat/sysinfo、
    设备/设备组、策略、RBAC、用户组）；M3 交付 96 个 operation
    （用户域 CRUD/邀请/批量/安全 14、通讯录 32、审计 9、仪表盘 2、
    服务器管理 10、nexus 9、设置 9、OIDC 提供者 8、更新检查 1、
    静态资源 2），总计 164 个 operation。
    错误响应统一 NestJS 包络 {statusCode, message, error}，
    message 三种形态并存（业务对象 / 纯文本 / 校验字符串数组）；
    POST 成功一律 200。限流 per-IP per-route（超限 429 ThrottlerException）；
    设备端协议为设备维度限流（tracker=id→uuid→IP）。
    契约基准：参考项目主分支实测（附录 A/B），仅取契约、不搬实现。
  version: 1.1.0-m2
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/permissions:
    get:
      summary: 权限目录只读（roles.view；{data:[...]} 无 total）
      description: 35+ 码编译期目录序列化，杜绝漂移；roles.create/edit/delete
        为 system_only（assignable=false）。
      operationId: listPermissions
      tags: [rbac]
      responses:
        "200":
          description: 权限目录
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PermissionDefinition"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Error"
  /api/permissions/me:
    get:
      summary: 当前用户生效权限（JWT；{permissions, scopes} 无 total）
      description: >-
        生效码经依赖过滤；scopes 形态：
        {global: [code...], device_group: {<guid>: [code...]}}。
      operationId: myPermissions
      tags: [rbac]
      responses:
        "200":
          description: 生效权限与 scope 映射
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EffectivePermissions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Error"
  /api/roles:
    get:
      summary: 角色列表（roles.view）
      description: name/note LIKE 过滤；name ASC、guid ASC 决胜排序。
      operationId: listRoles
      tags: [rbac]
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: note
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/CurrentParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: 角色分页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: 创建角色（super administrator）
      description: >-
        重名 409（Role name already exists）；system_only 码写入 400
        （Permission code cannot be assigned: ...）；requires 依赖缺失 400
        （Missing permission dependencies: ...）；事务内显式审计。
      operationId: createRole
      tags: [rbac]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateRequest"
      responses:
        "200":
          description: 创建后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /api/roles/{guid}:
    get:
      summary: 角色详情（roles.view；guid 需 uuid v4，否则 400）
      description: 含权限码清单。
      operationId: getRole
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 角色详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    patch:
      summary: 更新角色（super administrator）
      description: >-
        重名 409；取消保护（protected_account true→false）需
        confirm_protected_account_change=true，否则 400；
        system_only/requires 校验同创建；事务内显式审计。
      operationId: updateRole
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateRequest"
      responses:
        "200":
          description: 更新后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
    delete:
      summary: 删除角色（super administrator）
      description: >-
        事务级联删 role_permissions/user_role_assignments/
        user_role_assignment_device_groups；删除前 beforeState 全量快照审计。
      operationId: deleteRole
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 删除成功（空对象）
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/roles/{guid}/protection-impact:
    get:
      summary: 保护角色变更影响面（super administrator；{affected_member_count} 无 total）
      description: 统计挂在该角色上的用户数。
      operationId: roleProtectionImpact
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 影响面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtectionImpact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/users/{guid}/roles:
    get:
      summary: 用户角色指派列表（roles.assign；{data, effective_scope} 无 total）
      description: data 为 assignment 列表（createdAt ASC）；effective_scope 为
        聚合 scope（global | device_group | none）。
      operationId: getUserRoles
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 用户角色与聚合 scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRolesResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    put:
      summary: 全量替换用户角色指派（roles.assign；越权防护链）
      description: >-
        防护：自改角色 403（You cannot modify your own roles）；目标为保护账号
        需 super administrator（403）；给超管派非保护角色 403
        （Super administrators cannot be assigned regular roles）；
        device_group scope 仅限 device 档码（400）；device_group scope 须带 ≥1 组、
        global scope 不得带组（400）；非超管的组范围须在自己授权集内
        （403 Target device group is outside the authorized scope）。
      operationId: replaceUserRoles
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceRolesRequest"
      responses:
        "200":
          description: 替换后的 {data, effective_scope}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRolesResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /api/users/{guid}/roles/eligibility:
    get:
      summary: 角色指派资格矩阵（roles.assign；{data:[...]} 无 total）
      description: >-
        每个角色一行：eligible 与 reason_code（eligible / self_target /
        target_protected / protected_role / super_admin_target）。
      operationId: userRoleEligibility
      tags: [rbac]
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 资格矩阵
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/EligibilityRow"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"

  # ==================== M2 用户组域 ====================
...
components:
  parameters:
    CurrentParam:
      name: current
      in: query
      required: false
      description: 页码（1~100000，默认 1；camelCase 契约，禁止规范化）
      schema:
        type: integer
        minimum: 1
        maximum: 100000
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: 每页条数（1~100，默认 20；camelCase 契约，禁止规范化）
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    PermissionDefinition:
      type: object
      required: [code, resource, action, name, scope, assignable, system_only, requires]
      properties:
        code:
          type: string
        resource:
          type: string
        action:
          type: string
        name:
          type: string
        description:
          type: string
        scope:
          type: string
          enum: [global, device_group]
        assignable:
          type: boolean
        system_only:
          type: boolean
        requires:
          type: array
          items:
            type: string
    EffectivePermissions:
      type: object
      required: [permissions, scopes]
      properties:
        permissions:
          type: array
          items:
            type: string
          description: 生效码（依赖过滤后）
        scopes:
          type: object
          required: [global, device_group]
          properties:
            global:
              type: array
              items:
                type: string
            device_group:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              description: "<设备组guid>: [生效码...]"
    RoleView:
      type: object
      required: [guid, name, note, protected_account, created_at, updated_at]
      properties:
        guid:
          type: string
        name:
          type: string
        note:
          type: string
        protected_account:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RoleDetail:
      allOf:
        - $ref: "#/components/schemas/RoleView"
        - type: object
          required: [permissions]
          properties:
            permissions:
              type: array
              items:
                type: string
    RoleCreateRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
        note:
          type: string
        protected_account:
          type: boolean
        permissions:
          type: array
          items:
            type: string
    RoleUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        note:
          type: string
        protected_account:
          type: boolean
        permissions:
          type: array
          items:
            type: string
        confirm_protected_account_change:
          type: boolean
          description: 取消保护（true→false）必须为 true
    RolePage:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RoleView"
        total:
          type: integer
    ProtectionImpact:
      type: object
      required: [affected_member_count]
      properties:
        affected_member_count:
          type: integer
    AssignmentDto:
      type: object
      required: [role_guid, role_name, scope_type, device_group_guids]
      properties:
        role_guid:
          type: string
        role_name:
          type: string
        scope_type:
          type: string
          enum: [global, device_group]
        device_group_guids:
          type: array
          items:
            type: string
    UserRolesResult:
      type: object
      required: [data, effective_scope]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentDto"
        effective_scope:
          type: string
          enum: [global, device_group, none]
          description: 用户全部 assignment 聚合 scope
    ReplaceRolesRequest:
      type: object
      additionalProperties: false
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [role_guid, scope_type]
            properties:
              role_guid:
                type: string
              scope_type:
                type: string
                enum: [global, device_group]
              device_group_guids:
                type: array
                items:
                  type: string
    EligibilityRow:
      type: object
      required: [role_guid, role_name, protected_account, eligible, reason_code]
      properties:
        role_guid:
          type: string
        role_name:
          type: string
        protected_account:
          type: boolean
        eligible:
          type: boolean
        reason_code:
          type: string
          enum: [eligible, self_target, target_protected, protected_role, super_admin_target]
...
    Error:
      type: object
      description: NestJS 默认错误包络；message 为对象/字符串/字符串数组三种形态之一。
      required: [statusCode, message, error]
      properties:
        statusCode:
          type: integer
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
            - type: object
        error:
          type: string
  responses:
    BadRequest:
      description: 参数或校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证或会话已撤销
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Error:
      description: 错误包络
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 权限不足（RBAC scope / AdminGuard / super administrator）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 名称冲突（角色/用户组 409；设备组/策略为 400）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"